Compute boostrap statistics from a bootstrap run.

From a bootstrap run several statistics are computed. These are the model frequency, the bootstrap statistics of each parameter and, if the bootstrap is nonparametric, also the co-clustering probability of the input points. All the bootstrapped values are also returned. For the bootstrapped statistics the confidence intervals are returned for a required alpha value.

bootstrapped_statistics(x, bootstrap_results,
  bootstrap = "nonparametric", alpha = 0.025)

Arguments

x
A MOBSTER fit.

bootstrap_results
The results of running function mobster_bootstrap

bootstrap
Type of boostrap: "parametric" or "nonparametric"

alpha
Alpha-level empirical CIs for the bootstrap distribution.

Value

A list with the bootstrapped values, the model frequency, the bootstraped statistics and the co-clustering probability - non-null only for nonparametric boostrap.

Examples

# Random small dataset
dataset = random_dataset(N = 200, seed = 123, Beta_variance_scaling = 100)
x = mobster_fit(dataset$data, auto_setup = 'FAST')

#>  [ MOBSTER fit ] 
#> 

#> ✓ Loaded input data, n = 200.

#> ℹ n = 200. Mixture with k = 1,2 Beta(s). Pareto tail: TRUE and FALSE. Output clusters with π > 0.02 and n > 10.

#> ! mobster automatic setup FAST for the analysis.

#> ℹ Scoring (without parallel) 2 x 2 x 2 = 8 models by reICL.

#> 

#> [easypar] 2020-01-26 21:24:38 - Overriding parallel execution setup [FALSE] with global option : FALSE

#> |=================================                     | 62% ~1 s remaining     |========================================              | 75% ~1 s remaining     |===============================================       | 88% ~0 s remaining     |======================================================|100% ~0 s remaining     
#> 

#> ℹ MOBSTER best fit completed in 0.06 mins

#> 

#> ── [ MOBSTER ] My MOBSTER model n = 200 with k = 1 Beta(s) and a tail ─────────────────────────────────────────────────

#> ● Clusters: π = 85% [C1] and 15% [Tail], with π > 0.

#> ● Tail [n = 32, 15%] with alpha = 3.9.

#> ● Beta C1 [n = 168, 85%] with mean = 0.56.

#> ℹ Score(s): NLL = -41.48; ICL = -39.69 (-51.17), H = 11.48 (0). Fit converged by MM in 26 steps.

# Just 5 resamples of a nonparametric bootstrap run, disabling the parallel engine
options(easypar.parallel = FALSE)
boot_results = mobster_bootstrap(x$best, n.resamples = 5, auto_setup = 'FAST')

#>  [ MOBSTER bootstrap ~ 5 resamples from nonparametric bootstrap ] 
#> 

#> ── [ MOBSTER ] My MOBSTER model n = 200 with k = 1 Beta(s) and a tail ─────────────────────────────────────────────────

#> ● Clusters: π = 85% [C1] and 15% [Tail], with π > 0.

#> ● Tail [n = 32, 15%] with alpha = 3.9.

#> ● Beta C1 [n = 168, 85%] with mean = 0.56.

#> ℹ Score(s): NLL = -41.48; ICL = -39.69 (-51.17), H = 11.48 (0). Fit converged by MM in 26 steps.

#> 

#> 

#> ℹ Creating nonparametric bootstrap resamples

#>                                             

#> ✓ Creating nonparametric bootstrap resamples ... done

#> 

#> 

#> ── Running fits ─────────────────────────────────────────────────────────────────────────────── Might take some time ──

#> 

#> ℹ Fitting nonparametric bootstrap resamples

#> 

#> [easypar] 2020-01-26 21:24:42 - Overriding parallel execution setup [TRUE] with global option : FALSE

#>  [ MOBSTER fit ] 
#> 

#>                                            

#> ✓ Loaded input data, n = 200.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ℹ n = 200. Mixture with k = 1,2 Beta(s). Pareto tail: TRUE and FALSE. Output clusters with π > 0.02 and n > 10.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ! mobster automatic setup FAST for the analysis.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ℹ Scoring (without parallel) 2 x 2 x 2 = 8 models by reICL.

#> ℹ Fitting nonparametric bootstrap resamples

#> 

#> [easypar] 2020-01-26 21:24:42 - Overriding parallel execution setup [FALSE] with global option : FALSE

#> |===============================================       | 88% ~0 s remaining     |======================================================|100% ~0 s remaining     
#> 

#>                                            

#> ℹ MOBSTER best fit completed in 0.05 mins

#> ℹ Fitting nonparametric bootstrap resamples

#> 

#>                                            

#> ── [ MOBSTER ] My MOBSTER model n = 200 with k = 1 Beta(s) and a tail ─────────────────────────────────────────────────

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ● Clusters: π = 79% [C1] and 21% [Tail], with π > 0.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ● Tail [n = 43, 21%] with alpha = 3.1.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ● Beta C1 [n = 157, 79%] with mean = 0.55.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ℹ Score(s): NLL = -55.02; ICL = -65.88 (-78.24), H = 12.36 (0). Fit converged by MM in 18 steps.

#> ℹ Fitting nonparametric bootstrap resamples

#> |=====================                                 | 40% ~4 s remaining      [ MOBSTER fit ] 
#> 

#>                                            

#> ✓ Loaded input data, n = 200.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ℹ n = 200. Mixture with k = 1,2 Beta(s). Pareto tail: TRUE and FALSE. Output clusters with π > 0.02 and n > 10.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ! mobster automatic setup FAST for the analysis.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ℹ Scoring (without parallel) 2 x 2 x 2 = 8 models by reICL.

#> ℹ Fitting nonparametric bootstrap resamples

#> 

#> [easypar] 2020-01-26 21:24:45 - Overriding parallel execution setup [FALSE] with global option : FALSE

#> 
#> 

#>                                            

#> ℹ MOBSTER best fit completed in 0.04 mins

#> ℹ Fitting nonparametric bootstrap resamples

#> 

#>                                            

#> ── [ MOBSTER ] My MOBSTER model n = 200 with k = 2 Beta(s) and a tail ─────────────────────────────────────────────────

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ● Clusters: π = 45% [C2], 36% [C1] and 19% [Tail], with π > 0.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ● Tail [n = 37, 19%] with alpha = 4.2.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ● Beta C1 [n = 74, 36%] with mean = 0.81.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ● Beta C2 [n = 89, 45%] with mean = 0.4.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ℹ Score(s): NLL = -67.24; ICL = -69.8 (-75.1), H = 16.99 (11.68). Fit converged by MM in 68 steps.

#> ℹ Fitting nonparametric bootstrap resamples

#> |================================                      | 60% ~4 s remaining      [ MOBSTER fit ] 
#> 

#>                                            

#> ✓ Loaded input data, n = 200.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ℹ n = 200. Mixture with k = 1,2 Beta(s). Pareto tail: TRUE and FALSE. Output clusters with π > 0.02 and n > 10.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ! mobster automatic setup FAST for the analysis.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ℹ Scoring (without parallel) 2 x 2 x 2 = 8 models by reICL.

#> ℹ Fitting nonparametric bootstrap resamples

#> 

#> [easypar] 2020-01-26 21:24:47 - Overriding parallel execution setup [FALSE] with global option : FALSE

#> 
#> 

#>                                            

#> ℹ MOBSTER best fit completed in 0.04 mins

#> ℹ Fitting nonparametric bootstrap resamples

#> 

#>                                            

#> ── [ MOBSTER ] My MOBSTER model n = 200 with k = 2 Beta(s) and a tail ─────────────────────────────────────────────────

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ● Clusters: π = 51% [C2], 32% [C1] and 18% [Tail], with π > 0.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ● Tail [n = 35, 18%] with alpha = 4.3.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ● Beta C1 [n = 67, 32%] with mean = 0.81.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ● Beta C2 [n = 98, 51%] with mean = 0.42.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ℹ Score(s): NLL = -76.41; ICL = -90.9 (-94.45), H = 14.22 (10.68). Fit converged by MM in 78 steps.

#> ℹ Fitting nonparametric bootstrap resamples

#> |===========================================           | 80% ~2 s remaining      [ MOBSTER fit ] 
#> 

#>                                            

#> ✓ Loaded input data, n = 200.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ℹ n = 200. Mixture with k = 1,2 Beta(s). Pareto tail: TRUE and FALSE. Output clusters with π > 0.02 and n > 10.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ! mobster automatic setup FAST for the analysis.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ℹ Scoring (without parallel) 2 x 2 x 2 = 8 models by reICL.

#> ℹ Fitting nonparametric bootstrap resamples

#> 

#> [easypar] 2020-01-26 21:24:50 - Overriding parallel execution setup [FALSE] with global option : FALSE

#> 
#> 

#>                                            

#> ℹ MOBSTER best fit completed in 0.04 mins

#> ℹ Fitting nonparametric bootstrap resamples

#> 

#>                                            

#> ── [ MOBSTER ] My MOBSTER model n = 200 with k = 2 Beta(s) and a tail ─────────────────────────────────────────────────

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ● Clusters: π = 41% [C2], 37% [C1] and 22% [Tail], with π > 0.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ● Tail [n = 40, 22%] with alpha = 1.5.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ● Beta C1 [n = 74, 37%] with mean = 0.8.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ● Beta C2 [n = 86, 41%] with mean = 0.42.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ℹ Score(s): NLL = -46.19; ICL = -15.51 (-34.9), H = 29.19 (9.8). Fit converged by MM in 66 steps.

#> ℹ Fitting nonparametric bootstrap resamples

#> |======================================================|100% ~0 s remaining      [ MOBSTER fit ] 
#> 

#>                                            

#> ✓ Loaded input data, n = 200.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ℹ n = 200. Mixture with k = 1,2 Beta(s). Pareto tail: TRUE and FALSE. Output clusters with π > 0.02 and n > 10.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ! mobster automatic setup FAST for the analysis.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ℹ Scoring (without parallel) 2 x 2 x 2 = 8 models by reICL.

#> ℹ Fitting nonparametric bootstrap resamples

#> 

#> [easypar] 2020-01-26 21:24:53 - Overriding parallel execution setup [FALSE] with global option : FALSE

#> 
#> 

#>                                            

#> ℹ MOBSTER best fit completed in 0.04 mins

#> ℹ Fitting nonparametric bootstrap resamples

#> 

#>                                            

#> ── [ MOBSTER ] My MOBSTER model n = 200 with k = 1 Beta(s) and a tail ─────────────────────────────────────────────────

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ● Clusters: π = 87% [C1] and 13% [Tail], with π > 0.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ● Tail [n = 27, 13%] with alpha = 3.2.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ● Beta C1 [n = 173, 87%] with mean = 0.58.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ℹ Score(s): NLL = -37.54; ICL = -31.78 (-43.28), H = 11.5 (0). Fit converged by MM in 36 steps.

#> ℹ Fitting nonparametric bootstrap resamples

#>                                            

#> ✓ Fitting nonparametric bootstrap resamples ... done

#> 

boot_stats = bootstrapped_statistics(x$best, boot_results)

#> ℹ Bootstrap observations n = 5

#> 

#> ℹ Computing model frequency

#> # A tibble: 2 x 3
#>   Model           Frequency fit.model
#>   <fct>               <dbl> <lgl>    
#> 1 K = 2 with tail       0.6 FALSE    
#> 2 K = 1 with tail       0.4 TRUE     

#>                            

#> ✓ Computing model frequency ... done

#> 

#> 

#> ℹ Confidence Intervals (CI) for empirical quantiles

#> 
#> Mixing proportions 
#>  # A tibble: 3 x 8
#>   cluster statistics   min lower_quantile higher_quantile   max fit.value
#>   <chr>   <chr>      <dbl>          <dbl>           <dbl> <dbl>     <dbl>
#> 1 C1      Mixing pr… 0.319          0.323           0.862 0.870     0.854
#> 2 C2      Mixing pr… 0.412          0.414           0.503 0.505    NA    
#> 3 Tail    Mixing pr… 0.130          0.135           0.217 0.217     0.146
#> # … with 1 more variable: init.value <dbl>
#> 
#> Tail shape/ scale 
#>  # A tibble: 2 x 8
#>   cluster statistics    min lower_quantile higher_quantile    max fit.value
#>   <chr>   <chr>       <dbl>          <dbl>           <dbl>  <dbl>     <dbl>
#> 1 Tail    Scale      0.0504         0.0504          0.0532 0.0535    0.0504
#> 2 Tail    Shape      1.53           1.69            4.32   4.34      3.90  
#> # … with 1 more variable: init.value <dbl>
#> 
#> Beta peaks 
#>  # A tibble: 4 x 8
#>   cluster statistics     min lower_quantile higher_quantile    max fit.value
#>   <chr>   <chr>        <dbl>          <dbl>           <dbl>  <dbl>     <dbl>
#> 1 C1      Mean       0.552          0.555            0.808  0.808     0.558 
#> 2 C1      Variance   0.00464        0.00478          0.0487 0.0490    0.0524
#> 3 C2      Mean       0.395          0.396            0.420  0.420    NA     
#> 4 C2      Variance   0.00825        0.00857          0.0167 0.0168   NA     
#> # … with 1 more variable: init.value <dbl>

#>                                                    

#> ✓ Confidence Intervals (CI) for empirical quantiles ... done

#> 

#> 

#> ℹ Co-clustering probability from nonparametric bootstrap

#>                                                         

#> ✓ Co-clustering probability from nonparametric bootstrap ... done

#> 

print(boot_stats)

#> $bootstrap_values
#> # A tibble: 65 x 5
#>    cluster statistics        fit.value init.value resample
#>    <chr>   <chr>                 <dbl>      <dbl>    <int>
#>  1 Tail    Shape               3.14      3.54            1
#>  2 Tail    Scale               0.0504    0.0504          1
#>  3 Tail    Mean                0.0740    0.0703          1
#>  4 Tail    Variance            0.00154   0.000905        1
#>  5 C1      a                   2.44      7.12            1
#>  6 C1      b                   1.98      7.12            1
#>  7 C1      Mean                0.552     0.327           1
#>  8 C1      Variance            0.0456    0.00966         1
#>  9 Tail    Mixing proportion   0.208     0.5             1
#> 10 C1      Mixing proportion   0.792     0.5             1
#> # … with 55 more rows
#> 
#> $bootstrap_model
#> # A tibble: 2 x 3
#>   Model           Frequency fit.model
#>   <fct>               <dbl> <lgl>    
#> 1 K = 2 with tail       0.6 FALSE    
#> 2 K = 1 with tail       0.4 TRUE     
#> 
#> $bootstrap_statistics
#> # A tibble: 15 x 8
#>    cluster statistics     min lower_quantile higher_quantile      max fit.value
#>    <chr>   <chr>        <dbl>          <dbl>           <dbl>    <dbl>     <dbl>
#>  1 C1      a          2.30e+0       2.32             25.6     26.3     2.07    
#>  2 C1      b          1.66e+0       1.69              6.10     6.25    1.64    
#>  3 C1      Mean       5.52e-1       0.555             0.808    0.808   0.558   
#>  4 C1      Mixing pr… 3.19e-1       0.323             0.862    0.870   0.854   
#>  5 C1      Variance   4.64e-3       0.00478           0.0487   0.0490  0.0524  
#>  6 C2      a          5.23e+0       5.29             11.7     12.0    NA       
#>  7 C2      b          8.00e+0       8.05             16.2     16.6    NA       
#>  8 C2      Mean       3.95e-1       0.396             0.420    0.420  NA       
#>  9 C2      Mixing pr… 4.12e-1       0.414             0.503    0.505  NA       
#> 10 C2      Variance   8.25e-3       0.00857           0.0167   0.0168 NA       
#> 11 Tail    Mean       6.55e-2       0.0656            0.146    0.154   0.0678  
#> 12 Tail    Mixing pr… 1.30e-1       0.135             0.217    0.217   0.146   
#> 13 Tail    Scale      5.04e-2       0.0504            0.0532   0.0535  0.0504  
#> 14 Tail    Shape      1.53e+0       1.69              4.32     4.34    3.90    
#> 15 Tail    Variance   4.24e-4       0.000431        Inf      Inf       0.000622
#> # … with 1 more variable: init.value <dbl>
#> 
#> $bootstrap_co_clustering
#>     1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27 28
#> 1   0 2 2 3 2 2 3 1 2  3  2  2  3  1  2  3  2  2  1  1  1  0  2  1  2  2  2  1
#> 2   2 0 2 3 2 1 3 1 2  2  2  2  3  2  2  3  2  1  2  1  1  1  2  1  1  2  2  1
#> 3   2 2 0 3 3 2 3 2 1  2  1  3  3  1  2  3  1  2  1  0  0  1  3  0  2  3  2  2
#> 4   3 3 3 0 4 3 5 3 3  4  2  4  4  3  4  5  2  3  3  2  2  2  3  2  2  4  4  2
#> 5   2 2 3 4 0 3 4 3 2  3  1  4  3  2  3  4  1  3  2  1  1  2  3  1  2  4  3  2
#>     29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 50 51 52 53
#> 1    0  2  2  2  3  3  2  1  2  1  2  2  2  1  1  1  1  1  0  0  1  0  2  2  3
#> 2    1  1  3  1  3  2  2  2  1  2  1  3  2  2  1  2  2  2  1  1  1  1  2  1  3
#> 3    1  2  2  2  3  2  3  2  1  1  1  2  2  2  1  1  1  2  1  1  2  1  3  1  3
#> 4    2  3  3  3  4  3  3  3  3  3  3  4  3  3  2  3  2  3  2  1  2  2  4  2  4
#> 5    2  3  2  3  3  2  3  3  2  2  2  3  2  3  2  2  1  3  2  1  2  2  4  1  3
#>     54 55 56 57 58 59 60 61 62 63 64 65 66 67 68 69 70 71 72 73 74 75 76 77 78
#> 1    3  1  1  2  3  2  2  2  3  3  1  3  2  2  3  1  1  0  1  1  0  1  0  1  1
#> 2    2  1  2  2  3  2  1  1  2  3  2  2  1  2  3  2  1  0  1  1  0  1  0  1  1
#> 3    2  1  1  3  3  2  1  2  2  3  2  2  1  2  3  2  0  0  0  0  0  0  0  0  0
#> 4    3  2  2  4  5  4  3  2  4  5  2  4  3  4  4  3  2  1  2  1  0  2  0  2  1
#> 5    2  2  1  4  4  3  2  2  3  4  2  3  2  3  3  3  1  1  1  0  0  1  0  1  0
#>     79 80 81 82 83 84 85 86 87 88 89 90 91 92 93 94 95 96 97 98 99 100 101 102
#> 1    1  1  1  1  1  0  1  1  0  1  1  0  0  1  2  1  1  1  0  1  0   1   0   0
#> 2    1  1  1  1  1  0  1  1  0  1  1  0  0  1  2  1  1  1  0  1  0   1   0   0
#> 3    0  0  0  0  0  0  0  0  0  0  0  0  0  0  1  0  0  0  0  0  0   0   0   0
#> 4    1  2  2  2  2  1  2  2  1  1  2  1  1  2  3  1  2  2  1  1  1   2   1   1
#> 5    0  1  1  1  1  1  1  1  1  0  1  1  1  1  2  0  1  1  1  0  1   1   1   1
#>     103 104 105 106 107 108 109 110 111 112 113 114 115 116 117 118 119 120 121
#> 1     0   1   1   1   1   0   0   1   1   1   1   1   1   0   0   1   1   0   1
#> 2     0   1   1   1   1   0   0   1   1   1   1   1   1   0   0   1   1   0   1
#> 3     0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0
#> 4     1   2   2   1   2   0   1   1   1   1   1   1   2   1   1   2   2   0   2
#> 5     1   1   1   0   1   0   1   0   0   0   0   0   1   1   1   1   1   0   1
#>     122 123 124 125 126 127 128 129 130 131 132 133 134 135 136 137 138 139 140
#> 1     1   1   0   1   1   1   0   1   0   0   1   1   1   1   1   1   1   0   1
#> 2     1   1   0   1   1   1   0   1   0   0   1   1   1   1   1   1   1   0   1
#> 3     0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0
#> 4     1   1   1   2   2   1   0   1   0   1   2   2   2   2   1   1   2   0   2
#> 5     0   0   1   1   1   0   0   0   0   1   1   1   1   1   0   0   1   0   1
#>     141 142 143 144 145 146 147 148 149 152 153 156 157 166 170 171 172 175 176
#> 1     0   1   1   0   0   1   0   1   1   0   0   1   0   0   1   1   0   1   1
#> 2     0   1   1   0   0   1   0   1   1   0   0   1   0   0   1   1   0   1   1
#> 3     0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0
#> 4     0   2   1   0   1   2   0   2   2   0   1   1   1   1   2   2   1   1   1
#> 5     0   1   0   0   1   1   0   1   1   0   1   0   1   1   1   1   1   0   0
#>     178 179 180 188 191 192 195 196 199 150 151 154 155 158 159 160 161 162 163
#> 1     1   0   0   0   1   0   1   0   1   0   0   0   0   0   0   0   0   0   0
#> 2     1   0   0   0   1   0   1   0   1   0   0   0   0   0   0   0   0   0   0
#> 3     0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0
#> 4     1   0   1   0   2   0   2   0   2   0   0   0   0   0   0   0   0   0   0
#> 5     0   0   1   0   1   0   1   0   1   0   0   0   0   0   0   0   0   0   0
#>     164 165 167 168 169 173 174 177 181 182 183 184 185 186 187 189 190 193 194
#> 1     0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0
#> 2     0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0
#> 3     0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0
#> 4     0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0
#> 5     0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0   0
#>     197 198 200
#> 1     0   0   0
#> 2     0   0   0
#> 3     0   0   0
#> 4     0   0   0
#> 5     0   0   0
#>  [ reached getOption("max.print") -- omitted 195 rows ]
#>