Plot the boostrapped co-clustering probability.

From the results of a call to mobster_boostrap, and the results of a call to bootstrapped_statistics, a figure (heatmap-style) for the boostrapped co-clustering probability is produced assembling plots with cowplot::plot_grid.

plot_bootstrap_coclustering(
  x,
  bootstrap_results,
  bootstrap_statistics,
  colors = c(Tail = "gainsboro")
)

Arguments

x
A MOBSTER fit.

bootstrap_results
Results of a call to mobster_boostrap.

colors
If provided, these colours will be used for each cluster. If a subset of colours is provided, palette Set1 from RColorBrewer is used. By default the tail colour is provided as 'gainsboro'.

bootstrap_statisticsResults
of a call to bootstrapped_statistics.

Value

A cowplot figure of the bootstrapped co-clustering probability.

Examples

# Random small dataset
dataset = random_dataset(N = 200, seed = 123, Beta_variance_scaling = 100)
x = mobster_fit(dataset$data, auto_setup = 'FAST')

#>  [ MOBSTER fit ] 
#> 

#> Error in mobster:::check_input(x, K, samples, init, tail, epsilon, maxIter,     fit.type, seed, model.selection, trace): There are some reserved names in the input data that cannot be used, please remove or rename columns: cluster

# Just 5 resamples of a nonparametric bootstrap run, disabling the parallel engine
options(easypar.parallel = FALSE)
boot_results = mobster_bootstrap(x$best, n.resamples = 5, auto_setup = 'FAST')

#>  [ MOBSTER bootstrap ~ 5 resamples from nonparametric bootstrap ] 
#> 

#> Error in is_mobster_fit(x): object 'x' not found

boot_stats = bootstrapped_statistics(x$best, boot_results)

#> Error in is_mobster_fit(x): object 'x' not found

plot_bootstrap_coclustering(x, boot_results, boot_stats)

#> Error in "bootstrap" %in% names(x): object 'boot_results' not found