Plot the boostrapped mixing proportions

From the results of a call to mobster_boostrap, and the results of a call to bootstrapped_statistics, a boxplot with the mixing proportions is produced.

plot_bootstrap_mixing_proportions(x, bootstrap_results,
  bootstrap_statistics, colors = c(Tail = "gainsboro"))

Arguments

x
A MOBSTER fit.

bootstrap_results
Results of a call to mobster_boostrap.

colors
If provided, these colours will be used for each cluster. If a subset of colours is provided, palette Set1 from RColorBrewer is used. By default the tail colour is provided as 'gainsboro'.

bootstrap_statisticsResults
of a call to bootstrapped_statistics.

Value

A barplot of the bootstrapped mixing proportions.

Examples

# Random small dataset
dataset = random_dataset(N = 200, seed = 123, Beta_variance_scaling = 100)
x = mobster_fit(dataset$data, auto_setup = 'FAST')

#>  [ MOBSTER fit ] 
#> 

#> ✔ Loaded input data, n = 200.

#> ❯ n = 200. Mixture with k = 1,2 Beta(s). Pareto tail: TRUE and FALSE. Output clusters with π > 0.02 and n > 10.

#> ! mobster automatic setup FAST for the analysis.

#> ❯ Scoring (without parallel) 2 x 2 x 2 = 8 models by reICL.

#> 

#> [easypar] 2020-03-29 19:04:07 - Overriding parallel execution setup [FALSE] with global option : FALSE

#> 
#> 

#> ℹ MOBSTER fits completed in 2.8s.

#> 

#> ── [ MOBSTER ] My MOBSTER model n = 200 with k = 1 Beta(s) and a tail ───────────────────────────────────────────────────

#> ● Clusters: π = 85% [C1] and 15% [Tail], with π > 0.

#> ● Tail [n = 32, 15%] with alpha = 3.9.

#> ● Beta C1 [n = 168, 85%] with mean = 0.56.

#> ℹ Score(s): NLL = -41.48; ICL = -39.69 (-51.17), H = 11.48 (0). Fit converged by MM in 26 steps.

# Just 5 resamples of a nonparametric bootstrap run, disabling the parallel engine
options(easypar.parallel = FALSE)
boot_results = mobster_bootstrap(x$best, n.resamples = 5, auto_setup = 'FAST')

#>  [ MOBSTER bootstrap ~ 5 resamples from nonparametric bootstrap ] 
#> 

#> ── [ MOBSTER ] My MOBSTER model n = 200 with k = 1 Beta(s) and a tail ───────────────────────────────────────────────────

#> ● Clusters: π = 85% [C1] and 15% [Tail], with π > 0.

#> ● Tail [n = 32, 15%] with alpha = 3.9.

#> ● Beta C1 [n = 168, 85%] with mean = 0.56.

#> ℹ Score(s): NLL = -41.48; ICL = -39.69 (-51.17), H = 11.48 (0). Fit converged by MM in 26 steps.

#> 

#> 

#> ℹ Creating nonparametric bootstrap resamples

#>                                             

#> ✓ Creating nonparametric bootstrap resamples ... done

#> 

#> 

#> ── Running fits ──────────────────────────────────────────────────────────────────────────── Might take some time ...  ──

#> [easypar] 2020-03-29 19:04:10 - Overriding parallel execution setup [TRUE] with global option : FALSE

#>  [ MOBSTER fit ] 
#> 

#> ✔ Loaded input data, n = 200.

#> ❯ n = 200. Mixture with k = 1,2 Beta(s). Pareto tail: TRUE and FALSE. Output clusters with π > 0.02 and n > 10.

#> ! mobster automatic setup FAST for the analysis.

#> ❯ Scoring (without parallel) 2 x 2 x 2 = 8 models by reICL.

#> 

#> [easypar] 2020-03-29 19:04:10 - Overriding parallel execution setup [FALSE] with global option : FALSE

#> 
#> 

#> ℹ MOBSTER fits completed in 3.2s.

#> 

#> ── [ MOBSTER ] My MOBSTER model n = 200 with k = 1 Beta(s) and a tail ───────────────────────────────────────────────────

#> ● Clusters: π = 79% [C1] and 21% [Tail], with π > 0.

#> ● Tail [n = 43, 21%] with alpha = 3.1.

#> ● Beta C1 [n = 157, 79%] with mean = 0.55.

#> ℹ Score(s): NLL = -55.02; ICL = -65.88 (-78.24), H = 12.36 (0). Fit converged by MM in 18 steps.

#>  [ MOBSTER fit ] 
#> 

#> ✔ Loaded input data, n = 200.

#> ❯ n = 200. Mixture with k = 1,2 Beta(s). Pareto tail: TRUE and FALSE. Output clusters with π > 0.02 and n > 10.

#> ! mobster automatic setup FAST for the analysis.

#> ❯ Scoring (without parallel) 2 x 2 x 2 = 8 models by reICL.

#> 

#> [easypar] 2020-03-29 19:04:13 - Overriding parallel execution setup [FALSE] with global option : FALSE

#> 
#> 

#> ℹ MOBSTER fits completed in 3s.

#> 

#> ── [ MOBSTER ] My MOBSTER model n = 200 with k = 2 Beta(s) and a tail ───────────────────────────────────────────────────

#> ● Clusters: π = 45% [C2], 36% [C1] and 19% [Tail], with π > 0.

#> ● Tail [n = 37, 19%] with alpha = 4.2.

#> ● Beta C1 [n = 74, 36%] with mean = 0.81.

#> ● Beta C2 [n = 89, 45%] with mean = 0.4.

#> ℹ Score(s): NLL = -67.24; ICL = -69.8 (-75.1), H = 16.99 (11.68). Fit converged by MM in 68 steps.

#>  [ MOBSTER fit ] 
#> 

#> ✔ Loaded input data, n = 200.

#> ❯ n = 200. Mixture with k = 1,2 Beta(s). Pareto tail: TRUE and FALSE. Output clusters with π > 0.02 and n > 10.

#> ! mobster automatic setup FAST for the analysis.

#> ❯ Scoring (without parallel) 2 x 2 x 2 = 8 models by reICL.

#> 

#> [easypar] 2020-03-29 19:04:16 - Overriding parallel execution setup [FALSE] with global option : FALSE

#> 
#> 

#> ℹ MOBSTER fits completed in 2.8s.

#> 

#> ── [ MOBSTER ] My MOBSTER model n = 200 with k = 2 Beta(s) and a tail ───────────────────────────────────────────────────

#> ● Clusters: π = 51% [C2], 32% [C1] and 18% [Tail], with π > 0.

#> ● Tail [n = 35, 18%] with alpha = 4.3.

#> ● Beta C1 [n = 67, 32%] with mean = 0.81.

#> ● Beta C2 [n = 98, 51%] with mean = 0.42.

#> ℹ Score(s): NLL = -76.41; ICL = -90.9 (-94.45), H = 14.22 (10.68). Fit converged by MM in 78 steps.

#>  [ MOBSTER fit ] 
#> 

#> ✔ Loaded input data, n = 200.

#> ❯ n = 200. Mixture with k = 1,2 Beta(s). Pareto tail: TRUE and FALSE. Output clusters with π > 0.02 and n > 10.

#> ! mobster automatic setup FAST for the analysis.

#> ❯ Scoring (without parallel) 2 x 2 x 2 = 8 models by reICL.

#> 

#> [easypar] 2020-03-29 19:04:19 - Overriding parallel execution setup [FALSE] with global option : FALSE

#> 
#> 

#> ℹ MOBSTER fits completed in 2.8s.

#> 

#> ── [ MOBSTER ] My MOBSTER model n = 200 with k = 2 Beta(s) and a tail ───────────────────────────────────────────────────

#> ● Clusters: π = 41% [C2], 37% [C1] and 22% [Tail], with π > 0.

#> ● Tail [n = 40, 22%] with alpha = 1.5.

#> ● Beta C1 [n = 74, 37%] with mean = 0.8.

#> ● Beta C2 [n = 86, 41%] with mean = 0.42.

#> ℹ Score(s): NLL = -46.19; ICL = -15.51 (-34.9), H = 29.19 (9.8). Fit converged by MM in 66 steps.

#>  [ MOBSTER fit ] 
#> 

#> ✔ Loaded input data, n = 200.

#> ❯ n = 200. Mixture with k = 1,2 Beta(s). Pareto tail: TRUE and FALSE. Output clusters with π > 0.02 and n > 10.

#> ! mobster automatic setup FAST for the analysis.

#> ❯ Scoring (without parallel) 2 x 2 x 2 = 8 models by reICL.

#> 

#> [easypar] 2020-03-29 19:04:22 - Overriding parallel execution setup [FALSE] with global option : FALSE

#> 
#> 

#> ℹ MOBSTER fits completed in 2.7s.

#> 

#> ── [ MOBSTER ] My MOBSTER model n = 200 with k = 1 Beta(s) and a tail ───────────────────────────────────────────────────

#> ● Clusters: π = 87% [C1] and 13% [Tail], with π > 0.

#> ● Tail [n = 27, 13%] with alpha = 3.2.

#> ● Beta C1 [n = 173, 87%] with mean = 0.58.

#> ℹ Score(s): NLL = -37.54; ICL = -31.78 (-43.28), H = 11.5 (0). Fit converged by MM in 36 steps.

boot_stats = bootstrapped_statistics(x$best, boot_results)

#> 

#> ℹ Computing model frequency

#> # A tibble: 2 x 3
#>   Model           Frequency fit.model
#>   <fct>               <dbl> <lgl>    
#> 1 K = 2 with tail       0.6 FALSE    
#> 2 K = 1 with tail       0.4 TRUE     

#>                            

#> ✓ Computing model frequency ... done

#> 

#> 

#> ℹ Confidence Intervals (CI) for empirical quantiles

#> 
#> Mixing proportions 
#>  # A tibble: 3 x 8
#>   cluster statistics   min lower_quantile higher_quantile   max fit.value
#>   <chr>   <chr>      <dbl>          <dbl>           <dbl> <dbl>     <dbl>
#> 1 C1      Mixing pr… 0.319          0.323           0.862 0.870     0.854
#> 2 C2      Mixing pr… 0.412          0.414           0.503 0.505    NA    
#> 3 Tail    Mixing pr… 0.130          0.135           0.217 0.217     0.146
#> # … with 1 more variable: init.value <dbl>
#> 
#> Tail shape/ scale 
#>  # A tibble: 2 x 8
#>   cluster statistics    min lower_quantile higher_quantile    max fit.value
#>   <chr>   <chr>       <dbl>          <dbl>           <dbl>  <dbl>     <dbl>
#> 1 Tail    Scale      0.0504         0.0504          0.0532 0.0535    0.0504
#> 2 Tail    Shape      1.53           1.69            4.32   4.34      3.90  
#> # … with 1 more variable: init.value <dbl>
#> 
#> Beta peaks 
#>  # A tibble: 4 x 8
#>   cluster statistics     min lower_quantile higher_quantile    max fit.value
#>   <chr>   <chr>        <dbl>          <dbl>           <dbl>  <dbl>     <dbl>
#> 1 C1      Mean       0.552          0.555            0.808  0.808     0.558 
#> 2 C1      Variance   0.00464        0.00478          0.0487 0.0490    0.0524
#> 3 C2      Mean       0.395          0.396            0.420  0.420    NA     
#> 4 C2      Variance   0.00825        0.00857          0.0167 0.0168   NA     
#> # … with 1 more variable: init.value <dbl>

#>                                                    

#> ✓ Confidence Intervals (CI) for empirical quantiles ... done

#> 

#> 

#> ℹ Co-clustering probability from nonparametric bootstrap

#>                                                         

#> ✓ Co-clustering probability from nonparametric bootstrap ... done

#> 

plot_bootstrap_mixing_proportions(x, boot_results, boot_stats)

#> Error in UseMethod("filter_"): no applicable method for 'filter_' applied to an object of class "NULL"