Return the data with the hard clustering assigments.

Extract the clustering assignments, and return the data. Assignments can be computed subsetting the mutations that have posterior values (reponsibilities) below a certain cutoff (default 0 - all assigments); non-assigned mutations have NA as cluster label.

Clusters(x, cutoff_assignment = 0)

Arguments

x
A MOBSTER fit.

cutoff_assignment
Cutoff to compute hard clustering assignments.

Value

The data stored in x$data with a column label reporting the assigned cluster, or NA if the maximum cluster probability is below the threshold value cutoff_assignment.

Examples

data('fit_example', package = 'mobster')
Clusters(fit_example$best)

#> # A tibble: 426 x 7
#>    id       sample    VAF cluster   Tail         C1       C2
#>    <chr>    <chr>   <dbl> <chr>    <dbl>      <dbl>    <dbl>
#>  1 __mut1   S1     0.512  C2      0.0183 0.00000893 9.82e- 1
#>  2 __mut10  S1     0.505  C2      0.0184 0.0000140  9.82e- 1
#>  3 __mut100 S1     0.492  C2      0.0197 0.0000333  9.80e- 1
#>  4 __mut11  S1     0.447  C2      0.0450 0.000871   9.54e- 1
#>  5 __mut12  S1     0.515  C2      0.0185 0.00000699 9.82e- 1
#>  6 __mut125 S1     0.0505 Tail    1.000  0.00000162 1.53e-43
#>  7 __mut13  S1     0.539  C2      0.0226 0.00000166 9.77e- 1
#>  8 __mut14  S1     0.495  C2      0.0192 0.0000262  9.81e- 1
#>  9 __mut147 S1     0.299  C1      0.305  0.695      1.54e- 4
#> 10 __mut148 S1     0.206  C1      0.325  0.675      2.90e-11
#> # … with 416 more rows

# Add some cutoff to filter assignments
Clusters(fit_example$best, cutoff_assignment = .8)

#> # A tibble: 426 x 7
#>    id       sample    VAF cluster   Tail         C1       C2
#>    <chr>    <chr>   <dbl> <chr>    <dbl>      <dbl>    <dbl>
#>  1 __mut1   S1     0.512  C2      0.0183 0.00000893 9.82e- 1
#>  2 __mut10  S1     0.505  C2      0.0184 0.0000140  9.82e- 1
#>  3 __mut100 S1     0.492  C2      0.0197 0.0000333  9.80e- 1
#>  4 __mut11  S1     0.447  C2      0.0450 0.000871   9.54e- 1
#>  5 __mut12  S1     0.515  C2      0.0185 0.00000699 9.82e- 1
#>  6 __mut125 S1     0.0505 Tail    1.000  0.00000162 1.53e-43
#>  7 __mut13  S1     0.539  C2      0.0226 0.00000166 9.77e- 1
#>  8 __mut14  S1     0.495  C2      0.0192 0.0000262  9.81e- 1
#>  9 __mut147 S1     0.299  NA      0.305  0.695      1.54e- 4
#> 10 __mut148 S1     0.206  NA      0.325  0.675      2.90e-11
#> # … with 416 more rows