Compute boostrap statistics from a bootstrap run.

From a bootstrap run several statistics are computed. These are the model frequency, the bootstrap statistics of each parameter and, if the bootstrap is nonparametric, also the co-clustering probability of the input points. All the bootstrapped values are also returned. For the bootstrapped statistics the confidence intervals are returned for a required alpha value.

bootstrapped_statistics(
  x,
  bootstrap_results,
  bootstrap = "nonparametric",
  alpha = 0.025
)

Arguments

x
A MOBSTER fit.

bootstrap_results
The results of running function mobster_bootstrap

bootstrap
Type of boostrap: "parametric" or "nonparametric"

alpha
Alpha-level empirical CIs for the bootstrap distribution.

Value

A list with the bootstrapped values, the model frequency, the bootstraped statistics and the co-clustering probability - non-null only for nonparametric boostrap.

Examples

# Random small dataset
dataset = random_dataset(N = 200, seed = 123, Beta_variance_scaling = 100)

#> Error in data.frame(VAF = samples, cluster = ids, stringsAsFactors = FALSE) %>%     as_tibble(): could not find function "%>%"

x = mobster_fit(dataset$data, auto_setup = 'FAST')

#>  [ MOBSTER fit ] 
#> 

#> Error in stopifnot(is.matrix(x) | is.data.frame(x)): object 'dataset' not found

# Just 5 resamples of a nonparametric bootstrap run, disabling the parallel engine
options(easypar.parallel = FALSE)
boot_results = mobster_bootstrap(x$best, n.resamples = 5, auto_setup = 'FAST')

#>  [ MOBSTER bootstrap ~ 5 resamples from nonparametric bootstrap ] 
#> 

#> Error in is_mobster_fit(x): object 'x' not found

boot_stats = bootstrapped_statistics(x$best, boot_results)

#> Error in is_mobster_fit(x): object 'x' not found

print(boot_stats)

#> Error in print(boot_stats): object 'boot_stats' not found