Load data from a VCF file

This function parses a set of mutations in the Variant Calling Format (VCF), using the vcfR R package. The function takes in input the column names of the INFO field referring to allele depth counts, and number of reads with the variant. The two columns are used to compute the variant allele frequency (VAF) value; if any of the required columns are not found, an error is returned. This function does not filter any of the variants annotated in the input VCF file.

load_vcf(file, DP_column = "DP", NV_column = "NV")

Arguments

file
A VCF file name.

DP_column
Column with the total depth (DP), i.e., sum of reads at a locus.

NV_column
Column with the number of reads with the variant (NV).

Value

A tibble with the loaded data which contains, besides all the content parsable from the VCF file, three columns named DP, NV and VAF where VAF = NV/DP.

Examples

# Example VCF file in the https://github.com/openvax/varcode repository
file = 'https://raw.githubusercontent.com/openvax/varcode/master/test/data/strelka-example.vcf'

download.file(url = file, destfile = 'strelka-example.vcf')

load_vcf(file = 'strelka-example.vcf', DP_column)

#> ℹ Loading VCF file strelka-example.vcf with package vcfR, file size 11.6 Kb.

#> Error in DP_column %in% cn: object 'DP_column' not found