Bootstrap a MOBSTER fit.

Parametric and non-parametric implementation of bootstrap estimates for MOBSTER fits. This computation is parallel and uses ?easypar.

mobster_bootstrap(x, n.resamples = 100, bootstrap = "nonparametric",
  cores.ratio = 0.8, cache = NULL, save_data = NULL, ...)

Arguments

x
An object of class "dbpmm".

n.resamples
Number of boostrap resamples.

bootstrap
Type of boostrap: "parametric" or "nonparametric"

cores.ratio
Ratio of cores to use for the parallel; see ?easypar.

cache
Cache for the computation; see ?easypar.

...
fit parameters for mobster_fit

Value

Data from the fits, resamples and a plottable figure.

Examples

data('fit_example', package = 'mobster')

# Just 5 resamples of a nonparametric bootstrap run
mb = mobster_bootstrap(fit_example$best, n.resamples = 5)

#>  [ MOBSTER bootstrap ~ 5 resamples from nonparametric bootstrap ] 
#> 
#> =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
#>  Bootstrapping for this MOBSTER model
#> =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-

#> ── [ MOBSTER ] My MOBSTER model n = 5000, fit by MM in 75 steps (converged). ───────────────────────────────────────────

#> ● Tail [n = 1370, 31%] with alpha = 1.2.

#> ● Beta C1 [n = 2784, 55%] with mean = 0.48.

#> ● Beta C2 [n = 846, 14%] with mean = 0.15.

#> ℹ Score(s): NLL = -5671.5; ICL = -10359.09 (-11266.35), H = 907.26 (0), 

#> 
#> =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
#>  Creating nonparametric bootstrap resamples
#> =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
#> 
#> =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
#>  Running fits (might take some time)
#> =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-

# The resample data is wrapped as lists of lists
lapply(mb$resamples, function(x) x[[1]] %>% as_tibble)

#> [[1]]
#> # A tibble: 5,000 x 3
#>       id    VAF original.id
#>    <int>  <dbl>       <int>
#>  1     1 0.0740        4219
#>  2     2 0.451          581
#>  3     3 0.492         2443
#>  4     4 0.452         2402
#>  5     5 0.0867        3333
#>  6     6 0.475          885
#>  7     7 0.141         3608
#>  8     8 0.462         2092
#>  9     9 0.524         1269
#> 10    10 0.141         3289
#> # … with 4,990 more rows
#> 
#> [[2]]
#> # A tibble: 5,000 x 3
#>       id    VAF original.id
#>    <int>  <dbl>       <int>
#>  1     1 0.230         4041
#>  2     2 0.461          531
#>  3     3 0.129         4838
#>  4     4 0.467         1653
#>  5     5 0.0806        3574
#>  6     6 0.0797        4864
#>  7     7 0.0552        4658
#>  8     8 0.501         1435
#>  9     9 0.0515        4052
#> 10    10 0.0706        3360
#> # … with 4,990 more rows
#> 
#> [[3]]
#> # A tibble: 5,000 x 3
#>       id    VAF original.id
#>    <int>  <dbl>       <int>
#>  1     1 0.0505        3393
#>  2     2 0.479         2706
#>  3     3 0.498          864
#>  4     4 0.0600        3616
#>  5     5 0.150         4329
#>  6     6 0.463         1012
#>  7     7 0.438         2453
#>  8     8 0.262         3473
#>  9     9 0.464         2535
#> 10    10 0.141         2976
#> # … with 4,990 more rows
#> 
#> [[4]]
#> # A tibble: 5,000 x 3
#>       id    VAF original.id
#>    <int>  <dbl>       <int>
#>  1     1 0.465          545
#>  2     2 0.153         2919
#>  3     3 0.517         2193
#>  4     4 0.0591        3886
#>  5     5 0.463          912
#>  6     6 0.171         2893
#>  7     7 0.526         1373
#>  8     8 0.489          726
#>  9     9 0.117         3415
#> 10    10 0.158         2957
#> # … with 4,990 more rows
#> 
#> [[5]]
#> # A tibble: 5,000 x 3
#>       id    VAF original.id
#>    <int>  <dbl>       <int>
#>  1     1 0.501         2415
#>  2     2 0.554         3389
#>  3     3 0.448          954
#>  4     4 0.136         4679
#>  5     5 0.0505        3393
#>  6     6 0.521         1681
#>  7     7 0.466         2379
#>  8     8 0.451          644
#>  9     9 0.0679        4118
#> 10    10 0.0642        3701
#> # … with 4,990 more rows
#> 

# The best fits are returned
mb$fits

#> $`1`

#> ── [ MOBSTER ] My MOBSTER model n = 5000, fit by MM in 19 steps (converged). ───────────────────────────────────────────

#> ● Tail [n = 2217, 46%] with alpha = 1.1.

#> ● Beta C1 [n = 2783, 54%] with mean = 0.48.

#> ℹ Score(s): NLL = -5324.75; ICL = -10271.07 (-10598.4), H = 327.33 (0), 

#> 
#> $`2`

#> ── [ MOBSTER ] My MOBSTER model n = 5000, fit by MM in 18 steps (converged). ───────────────────────────────────────────

#> ● Tail [n = 2200, 45%] with alpha = 1.1.

#> ● Beta C1 [n = 2800, 55%] with mean = 0.48.

#> ℹ Score(s): NLL = -5241.31; ICL = -10099.69 (-10431.51), H = 331.82 (0), 

#> 
#> $`3`

#> ── [ MOBSTER ] My MOBSTER model n = 5000, fit by MM in 15 steps (converged). ───────────────────────────────────────────

#> ● Tail [n = 2245, 46%] with alpha = 1.1.

#> ● Beta C1 [n = 2755, 54%] with mean = 0.48.

#> ℹ Score(s): NLL = -5350.93; ICL = -10331.16 (-10650.75), H = 319.59 (0), 

#> 
#> $`4`

#> ── [ MOBSTER ] My MOBSTER model n = 5000, fit by MM in 187 steps (converged). ──────────────────────────────────────────

#> ● Tail [n = 1389, 32%] with alpha = 1.2.

#> ● Beta C1 [n = 2760, 54%] with mean = 0.48.

#> ● Beta C2 [n = 851, 14%] with mean = 0.15.

#> ℹ Score(s): NLL = -5531.05; ICL = -10066.82 (-10985.44), H = 918.62 (0), 

#> 
#> $`5`

#> ── [ MOBSTER ] My MOBSTER model n = 5000, fit by MM in 125 steps (converged). ──────────────────────────────────────────

#> ● Tail [n = 1348, 31%] with alpha = 1.2.

#> ● Beta C1 [n = 2793, 55%] with mean = 0.48.

#> ● Beta C2 [n = 859, 14%] with mean = 0.15.

#> ℹ Score(s): NLL = -5715.9; ICL = -10443.21 (-11355.14), H = 911.93 (0), 

#>