Bootstrap a MOBSTER fit.

Parametric and non-parametric implementation of bootstrap estimates for MOBSTER fits. This computation is parallel and uses ?easypar.

mobster_bootstrap(x, n.resamples = 100, bootstrap = "nonparametric",
  cores.ratio = 0.8, cache = NULL, save_data = NULL, ...)

Arguments

x
An object of class "dbpmm".

n.resamples
Number of boostrap resamples.

bootstrap
Type of boostrap: "parametric" or "nonparametric"

cores.ratio
Ratio of cores to use for the parallel; see ?easypar.

cache
Cache for the computation; see ?easypar.

...
fit parameters for mobster_fit

Value

Data from the fits, resamples and a plottable figure.

Examples

# Random small dataset
dataset = random_dataset(N = 200, seed = 123, Beta_variance_scaling = 100)
x = mobster_fit(dataset$data, auto_setup = 'FAST')

#>  [ MOBSTER fit ] 
#> 

#> ✓ Loaded input data, n = 200.

#> ℹ n = 200. Mixture with k = 1,2 Beta(s). Pareto tail: TRUE and FALSE. Output clusters with π > 0.02 and n > 10.

#> ! mobster automatic setup FAST for the analysis.

#> ℹ Scoring (without parallel) 2 x 2 x 2 = 8 models by reICL.

#> 

#> [easypar] 2020-02-08 18:18:24 - Overriding parallel execution setup [FALSE] with global option : FALSE

#> 
#> 

#> ℹ MOBSTER best fit completed in 0.06 mins

#> 

#> ── [ MOBSTER ] My MOBSTER model n = 200 with k = 1 Beta(s) and a tail ──────────────────────────────────

#> ● Clusters: π = 85% [C1] and 15% [Tail], with π > 0.

#> ● Tail [n = 32, 15%] with alpha = 3.9.

#> ● Beta C1 [n = 168, 85%] with mean = 0.56.

#> ℹ Score(s): NLL = -41.48; ICL = -39.69 (-51.17), H = 11.48 (0). Fit converged by MM in 26 steps.

# Just 5 resamples of a nonparametric bootstrap run, disabling the parallel engine
options(easypar.parallel = FALSE)
boot_results = mobster_bootstrap(x$best, n.resamples = 5, auto_setup = 'FAST')

#>  [ MOBSTER bootstrap ~ 5 resamples from nonparametric bootstrap ] 
#> 

#> ── [ MOBSTER ] My MOBSTER model n = 200 with k = 1 Beta(s) and a tail ──────────────────────────────────

#> ● Clusters: π = 85% [C1] and 15% [Tail], with π > 0.

#> ● Tail [n = 32, 15%] with alpha = 3.9.

#> ● Beta C1 [n = 168, 85%] with mean = 0.56.

#> ℹ Score(s): NLL = -41.48; ICL = -39.69 (-51.17), H = 11.48 (0). Fit converged by MM in 26 steps.

#> 

#> 

#> ℹ Creating nonparametric bootstrap resamples

#>                                             

#> ✓ Creating nonparametric bootstrap resamples ... done

#> 

#> 

#> ── Running fits ──────────────────────────────────────────────────────────────── Might take some time ──

#> [easypar] 2020-02-08 18:18:28 - Overriding parallel execution setup [TRUE] with global option : FALSE

#>  [ MOBSTER fit ] 
#> 

#> ✓ Loaded input data, n = 200.

#> ℹ n = 200. Mixture with k = 1,2 Beta(s). Pareto tail: TRUE and FALSE. Output clusters with π > 0.02 and n > 10.

#> ! mobster automatic setup FAST for the analysis.

#> ℹ Scoring (without parallel) 2 x 2 x 2 = 8 models by reICL.

#> 

#> [easypar] 2020-02-08 18:18:28 - Overriding parallel execution setup [FALSE] with global option : FALSE

#> 
#> 

#> ℹ MOBSTER best fit completed in 0.05 mins

#> 

#> ── [ MOBSTER ] My MOBSTER model n = 200 with k = 1 Beta(s) and a tail ──────────────────────────────────

#> ● Clusters: π = 79% [C1] and 21% [Tail], with π > 0.

#> ● Tail [n = 43, 21%] with alpha = 3.1.

#> ● Beta C1 [n = 157, 79%] with mean = 0.55.

#> ℹ Score(s): NLL = -55.02; ICL = -65.88 (-78.24), H = 12.36 (0). Fit converged by MM in 18 steps.

#>  [ MOBSTER fit ] 
#> 

#> ✓ Loaded input data, n = 200.

#> ℹ n = 200. Mixture with k = 1,2 Beta(s). Pareto tail: TRUE and FALSE. Output clusters with π > 0.02 and n > 10.

#> ! mobster automatic setup FAST for the analysis.

#> ℹ Scoring (without parallel) 2 x 2 x 2 = 8 models by reICL.

#> 

#> [easypar] 2020-02-08 18:18:31 - Overriding parallel execution setup [FALSE] with global option : FALSE

#> 
#> 

#> ℹ MOBSTER best fit completed in 0.05 mins

#> 

#> ── [ MOBSTER ] My MOBSTER model n = 200 with k = 2 Beta(s) and a tail ──────────────────────────────────

#> ● Clusters: π = 45% [C2], 36% [C1] and 19% [Tail], with π > 0.

#> ● Tail [n = 37, 19%] with alpha = 4.2.

#> ● Beta C1 [n = 74, 36%] with mean = 0.81.

#> ● Beta C2 [n = 89, 45%] with mean = 0.4.

#> ℹ Score(s): NLL = -67.24; ICL = -69.8 (-75.1), H = 16.99 (11.68). Fit converged by MM in 68 steps.

#>  [ MOBSTER fit ] 
#> 

#> ✓ Loaded input data, n = 200.

#> ℹ n = 200. Mixture with k = 1,2 Beta(s). Pareto tail: TRUE and FALSE. Output clusters with π > 0.02 and n > 10.

#> ! mobster automatic setup FAST for the analysis.

#> ℹ Scoring (without parallel) 2 x 2 x 2 = 8 models by reICL.

#> 

#> [easypar] 2020-02-08 18:18:34 - Overriding parallel execution setup [FALSE] with global option : FALSE

#> 
#> 

#> ℹ MOBSTER best fit completed in 0.05 mins

#> 

#> ── [ MOBSTER ] My MOBSTER model n = 200 with k = 2 Beta(s) and a tail ──────────────────────────────────

#> ● Clusters: π = 51% [C2], 32% [C1] and 18% [Tail], with π > 0.

#> ● Tail [n = 35, 18%] with alpha = 4.3.

#> ● Beta C1 [n = 67, 32%] with mean = 0.81.

#> ● Beta C2 [n = 98, 51%] with mean = 0.42.

#> ℹ Score(s): NLL = -76.41; ICL = -90.9 (-94.45), H = 14.22 (10.68). Fit converged by MM in 78 steps.

#>  [ MOBSTER fit ] 
#> 

#> ✓ Loaded input data, n = 200.

#> ℹ n = 200. Mixture with k = 1,2 Beta(s). Pareto tail: TRUE and FALSE. Output clusters with π > 0.02 and n > 10.

#> ! mobster automatic setup FAST for the analysis.

#> ℹ Scoring (without parallel) 2 x 2 x 2 = 8 models by reICL.

#> 

#> [easypar] 2020-02-08 18:18:37 - Overriding parallel execution setup [FALSE] with global option : FALSE

#> 
#> 

#> ℹ MOBSTER best fit completed in 0.06 mins

#> 

#> ── [ MOBSTER ] My MOBSTER model n = 200 with k = 2 Beta(s) and a tail ──────────────────────────────────

#> ● Clusters: π = 41% [C2], 37% [C1] and 22% [Tail], with π > 0.

#> ● Tail [n = 40, 22%] with alpha = 1.5.

#> ● Beta C1 [n = 74, 37%] with mean = 0.8.

#> ● Beta C2 [n = 86, 41%] with mean = 0.42.

#> ℹ Score(s): NLL = -46.19; ICL = -15.51 (-34.9), H = 29.19 (9.8). Fit converged by MM in 66 steps.

#>  [ MOBSTER fit ] 
#> 

#> ✓ Loaded input data, n = 200.

#> ℹ n = 200. Mixture with k = 1,2 Beta(s). Pareto tail: TRUE and FALSE. Output clusters with π > 0.02 and n > 10.

#> ! mobster automatic setup FAST for the analysis.

#> ℹ Scoring (without parallel) 2 x 2 x 2 = 8 models by reICL.

#> 

#> [easypar] 2020-02-08 18:18:41 - Overriding parallel execution setup [FALSE] with global option : FALSE

#> 
#> 

#> ℹ MOBSTER best fit completed in 0.05 mins

#> 

#> ── [ MOBSTER ] My MOBSTER model n = 200 with k = 1 Beta(s) and a tail ──────────────────────────────────

#> ● Clusters: π = 87% [C1] and 13% [Tail], with π > 0.

#> ● Tail [n = 27, 13%] with alpha = 3.2.

#> ● Beta C1 [n = 173, 87%] with mean = 0.58.

#> ℹ Score(s): NLL = -37.54; ICL = -31.78 (-43.28), H = 11.5 (0). Fit converged by MM in 36 steps.

# The resample data is available in a list
print(boot_results$resamples[[1]])

#> [[1]]
#>      id        VAF original.id
#> 1     1 0.71161969          63
#> 2     2 0.67803199          62
#> 3     3 0.56979786         105
#> 4     4 0.40933369         110
#> 5     5 0.05579153         200
#> 6     6 0.71525460          56
#> 7     7 0.79487310          37
#> 8     8 0.88075926          11
#> 9     9 0.41345019         137
#> 10   10 0.06459001         164
#> 11   11 0.75981874          10
#> 12   12 0.16891439         199
#> 13   13 0.88008176          58
#> 14   14 0.40287024          77
#> 15   15 0.45989376          98
#> 16   16 0.40855891          82
#> 17   17 0.05549647         174
#> 18   18 0.53971624         129
#> 19   19 0.22631181         178
#> 20   20 0.56979786         105
#> 21   21 0.05041418         183
#> 22   22 0.40855891          82
#> 23   23 0.62660125          93
#> 24   24 0.05292696         189
#> 25   25 0.07101218         197
#> 26   26 0.66991625          13
#> 27   27 0.24854800         175
#> 28   28 0.24854800         175
#> 29   29 0.35323852          79
#> 30   30 0.09339962         187
#> 31   31 0.67803199          62
#> 32   32 0.05307622         161
#> 33   33 0.39715110         119
#> 34   34 0.07749332         190
#> 35   35 0.41083410         122
#> 36   36 0.07101218         197
#> 37   37 0.87201302          14
#> 38   38 0.34937041         195
#> 39   39 0.88900588          31
#> 40   40 0.51906602         136
#> 41   41 0.34532994         111
#> 42   42 0.05549647         174
#> 43   43 0.48504929          86
#> 44   44 0.73728252          60
#> 45   45 0.67803199          62
#> 46   46 0.68272022          66
#> 47   47 0.93508454           1
#> 48   48 0.45889813         113
#> 49   49 0.48504929          86
#> 50   50 0.87223851          38
#> 51   51 0.49957893          88
#> 52   52 0.71525460          56
#> 53   53 0.42237360         106
#> 54   54 0.29071784         143
#> 55   55 0.46922452         138
#> 56   56 0.40855891          82
#> 57   57 0.31918073         123
#> 58   58 0.54149138          94
#> 59   59 0.31918073         123
#> 60   60 0.52507675          81
#> 61   61 0.79626126           4
#> 62   62 0.05560013         168
#> 63   63 0.41083410         122
#> 64   64 0.90938604          34
#> 65   65 0.23328957         100
#> 66   66 0.34937041         195
#> 67   67 0.40933369         110
#> 68   68 0.71161969          63
#> 69   69 0.73852333          67
#> 70   70 0.85119549          59
#> 71   71 0.77727140          45
#> 72   72 0.09848206         185
#> 73   73 0.41344594          95
#> 74   74 0.46649786         140
#> 75   75 0.08286166         159
#> 76   76 0.94033592          40
#> 77   77 0.51906602         136
#> 78   78 0.60215090          24
#> 79   79 0.25356548         191
#> 80   80 0.41345019         137
#> 81   81 0.06451491         160
#> 82   82 0.84183709          15
#> 83   83 0.41820206         170
#> 84   84 0.68272022          66
#> 85   85 0.80575738          20
#> 86   86 0.08286166         159
#> 87   87 0.07101218         197
#> 88   88 0.09371586         154
#> 89   89 0.05982373         151
#> 90   90 0.85119549          59
#> 91   91 0.36479148          96
#> 92   92 0.35389360         148
#> 93   93 0.14237637         156
#> 94   94 0.69855219          16
#> 95   95 0.88900588          31
#> 96   96 0.50441175         125
#> 97   97 0.68272022          66
#> 98   98 0.27063711          83
#> 99   99 0.51489175          72
#> 100 100 0.05288930         165
#> 101 101 0.07739772         194
#> 102 102 0.20345665         171
#> 103 103 0.86751858          21
#> 104 104 0.06459001         164
#> 105 105 0.49679960         132
#> 106 106 0.35389360         148
#> 107 107 0.50353374         134
#> 108 108 0.34532994         111
#> 109 109 0.84477267          27
#> 110 110 0.73227401          65
#> 111 111 0.41146463         127
#> 112 112 0.41591364          78
#> 113 113 0.42106068         135
#> 114 114 0.38618946          92
#> 115 115 0.06459001         164
#> 116 116 0.94987856          17
#> 117 117 0.80708263          33
#> 118 118 0.05047598         177
#> 119 119 0.05928644         182
#> 120 120 0.41344594          95
#> 121 121 0.44468643         149
#> 122 122 0.34937041         195
#> 123 123 0.42106068         135
#> 124 124 0.83160360          54
#> 125 125 0.84183709          15
#> 126 126 0.44468643         149
#> 127 127 0.42106068         135
#> 128 128 0.40898455          73
#> 129 129 0.06559523         193
#> 130 130 0.88008176          58
#> 131 131 0.54678488         126
#> 132 132 0.24854800         175
#> 133 133 0.55706664         104
#> 134 134 0.49193870         146
#> 135 135 0.05794455         158
#> 136 136 0.53065152          89
#> 137 137 0.35553257          75
#> 138 138 0.39313204          85
#> 139 139 0.48504929          86
#> 140 140 0.09339962         187
#> 141 141 0.88008176          58
#> 142 142 0.30218684          70
#> 143 143 0.31918073         123
#> 144 144 0.27063711          83
#> 145 145 0.60215090          24
#> 146 146 0.60215090          24
#> 147 147 0.09632256         181
#> 148 148 0.08276675         162
#> 149 149 0.86311298           7
#> 150 150 0.40287024          77
#> 151 151 0.50929823          80
#> 152 152 0.83847963          39
#> 153 153 0.48504929          86
#> 154 154 0.74356586          53
#> 155 155 0.08276675         162
#> 156 156 0.05307622         161
#> 157 157 0.82221396           9
#> 158 158 0.06451491         160
#> 159 159 0.36961332         133
#> 160 160 0.05388263         150
#> 161 161 0.05579153         200
#> 162 162 0.88574939          52
#> 163 163 0.51489175          72
#> 164 164 0.89633905          68
#> 165 165 0.20345665         171
#> 166 166 0.71525460          56
#> 167 167 0.07119888         167
#> 168 168 0.08191177         186
#> 169 169 0.48474986         118
#> 170 170 0.84727835           2
#> 171 171 0.79487310          37
#> 172 172 0.75981874          10
#> 173 173 0.47953999         112
#> 174 174 0.38618946          92
#> 175 175 0.68272022          66
#> 176 176 0.27063711          83
#> 177 177 0.33158809         114
#> 178 178 0.05292696         189
#> 179 179 0.83331149          19
#> 180 180 0.05307622         161
#> 181 181 0.23209963         176
#> 182 182 0.69855219          16
#> 183 183 0.60215090          24
#> 184 184 0.08191177         186
#> 185 185 0.46922452         138
#> 186 186 0.45889813         113
#> 187 187 0.45523837         142
#> 188 188 0.09371586         154
#> 189 189 0.75322960          41
#> 190 190 0.62660125          93
#> 191 191 0.82224188          44
#> 192 192 0.38618946          92
#> 193 193 0.55465880         115
#> 194 194 0.48504929          86
#> 195 195 0.08286166         159
#> 196 196 0.82221396           9
#> 197 197 0.49091509         107
#> 198 198 0.85119549          59
#> 199 199 0.52803851         121
#> 200 200 0.09632256         181
#> 

# The best fits are returned
print(boot_results$fits)

#> $`1`

#> ── [ MOBSTER ] My MOBSTER model n = 200 with k = 1 Beta(s) and a tail ──────────────────────────────────

#> ● Clusters: π = 79% [C1] and 21% [Tail], with π > 0.

#> ● Tail [n = 43, 21%] with alpha = 3.1.

#> ● Beta C1 [n = 157, 79%] with mean = 0.55.

#> ℹ Score(s): NLL = -55.02; ICL = -65.88 (-78.24), H = 12.36 (0). Fit converged by MM in 18 steps.

#> 
#> $`2`

#> ── [ MOBSTER ] My MOBSTER model n = 200 with k = 2 Beta(s) and a tail ──────────────────────────────────

#> ● Clusters: π = 45% [C2], 36% [C1] and 19% [Tail], with π > 0.

#> ● Tail [n = 37, 19%] with alpha = 4.2.

#> ● Beta C1 [n = 74, 36%] with mean = 0.81.

#> ● Beta C2 [n = 89, 45%] with mean = 0.4.

#> ℹ Score(s): NLL = -67.24; ICL = -69.8 (-75.1), H = 16.99 (11.68). Fit converged by MM in 68 steps.

#> 
#> $`3`

#> ── [ MOBSTER ] My MOBSTER model n = 200 with k = 2 Beta(s) and a tail ──────────────────────────────────

#> ● Clusters: π = 51% [C2], 32% [C1] and 18% [Tail], with π > 0.

#> ● Tail [n = 35, 18%] with alpha = 4.3.

#> ● Beta C1 [n = 67, 32%] with mean = 0.81.

#> ● Beta C2 [n = 98, 51%] with mean = 0.42.

#> ℹ Score(s): NLL = -76.41; ICL = -90.9 (-94.45), H = 14.22 (10.68). Fit converged by MM in 78 steps.

#> 
#> $`4`

#> ── [ MOBSTER ] My MOBSTER model n = 200 with k = 2 Beta(s) and a tail ──────────────────────────────────

#> ● Clusters: π = 41% [C2], 37% [C1] and 22% [Tail], with π > 0.

#> ● Tail [n = 40, 22%] with alpha = 1.5.

#> ● Beta C1 [n = 74, 37%] with mean = 0.8.

#> ● Beta C2 [n = 86, 41%] with mean = 0.42.

#> ℹ Score(s): NLL = -46.19; ICL = -15.51 (-34.9), H = 29.19 (9.8). Fit converged by MM in 66 steps.

#> 
#> $`5`

#> ── [ MOBSTER ] My MOBSTER model n = 200 with k = 1 Beta(s) and a tail ──────────────────────────────────

#> ● Clusters: π = 87% [C1] and 13% [Tail], with π > 0.

#> ● Tail [n = 27, 13%] with alpha = 3.2.

#> ● Beta C1 [n = 173, 87%] with mean = 0.58.

#> ℹ Score(s): NLL = -37.54; ICL = -31.78 (-43.28), H = 11.5 (0). Fit converged by MM in 36 steps.

#>