Bootstrap a MOBSTER fit.

Parametric and non-parametric implementation of bootstrap estimates for MOBSTER fits. This computation is parallel and uses ?easypar.

mobster_bootstrap(x, n.resamples = 100, bootstrap = "nonparametric",
  cores.ratio = 0.8, cache = NULL, save_data = NULL, ...)

Arguments

x
An object of class "dbpmm".

n.resamples
Number of boostrap resamples.

bootstrap
Type of boostrap: "parametric" or "nonparametric"

cores.ratio
Ratio of cores to use for the parallel; see ?easypar.

cache
Cache for the computation; see ?easypar.

...
fit parameters for mobster_fit

Value

Data from the fits, resamples and a plottable figure.

Examples

data('fit_example', package = 'mobster')

# Just 5 resamples of a nonparametric bootstrap run
mb = mobster_bootstrap(fit_example$best, n.resamples = 5)

#>  [ MOBSTER bootstrap ~ 5 resamples from nonparametric bootstrap ] 
#> 
#> =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
#>  Bootstrapping for this MOBSTER model
#> =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
#> [ MOBSTER ] N = 426 points with K = 2 Beta clusters, fit by MM in 289 steps [CONVERGED] 
#> 
#>   Components (fit)  
#>      Tail	n = 259 (0.64) 	 Shape = 1.8 
#>   Beta C1 	n = 63 (0.13) 	 Mean = 0.23
#>   Beta C2 	n = 104 (0.24) 	 Mean = 0.5
#> 
#>   Scores (model selection)  
#>        NLL       BIC       AIC  entropy       ICL     reICL size
#>  -490.8661 -909.0789 -957.7322 80.79651 -828.2824 -908.2693   12
#> 
#> =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
#>  Creating nonparametric bootstrap resamples
#> =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
#> 
#> =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
#>  Running fits (might take some time)
#> =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-

# The resample data is wrapped as lists of lists
lapply(mb$resamples, function(x) x[[1]] %>% as_tibble)

#> [[1]]
#> # A tibble: 426 x 3
#>       id    VAF original.id
#>    <int>  <dbl>       <int>
#>  1     1 0.0833         263
#>  2     2 0.371          356
#>  3     3 0.0722         255
#>  4     4 0.0891         309
#>  5     5 0.514          338
#>  6     6 0.0595         343
#>  7     7 0.0575          23
#>  8     8 0.435          328
#>  9     9 0.0990          97
#> 10    10 0.510          322
#> # … with 416 more rows
#> 
#> [[2]]
#> # A tibble: 426 x 3
#>       id    VAF original.id
#>    <int>  <dbl>       <int>
#>  1     1 0.115          115
#>  2     2 0.184          138
#>  3     3 0.135          249
#>  4     4 0.0722         104
#>  5     5 0.156          382
#>  6     6 0.0571          66
#>  7     7 0.205           34
#>  8     8 0.0505         284
#>  9     9 0.152          302
#> 10    10 0.299            9
#> # … with 416 more rows
#> 
#> [[3]]
#> # A tibble: 426 x 3
#>       id    VAF original.id
#>    <int>  <dbl>       <int>
#>  1     1 0.0962         106
#>  2     2 0.546          425
#>  3     3 0.125          148
#>  4     4 0.0625         334
#>  5     5 0.546          425
#>  6     6 0.5            406
#>  7     7 0.0581         208
#>  8     8 0.278           16
#>  9     9 0.112          394
#> 10    10 0.546          113
#> # … with 416 more rows
#> 
#> [[4]]
#> # A tibble: 426 x 3
#>       id    VAF original.id
#>    <int>  <dbl>       <int>
#>  1     1 0.0571          66
#>  2     2 0.556          254
#>  3     3 0.0988         111
#>  4     4 0.0611         311
#>  5     5 0.0505         284
#>  6     6 0.104          215
#>  7     7 0.0744         380
#>  8     8 0.539            7
#>  9     9 0.0722         255
#> 10    10 0.0686         314
#> # … with 416 more rows
#> 
#> [[5]]
#> # A tibble: 426 x 3
#>       id    VAF original.id
#>    <int>  <dbl>       <int>
#>  1     1 0.237          149
#>  2     2 0.478           17
#>  3     3 0.0686         286
#>  4     4 0.0795         275
#>  5     5 0.529          408
#>  6     6 0.0638         130
#>  7     7 0.0860         395
#>  8     8 0.131          371
#>  9     9 0.0510         342
#> 10    10 0.419          357
#> # … with 416 more rows
#> 

# The best fits are returned
mb$fits

#> $`1`
#> [ MOBSTER ] N = 426 points with K = 2 Beta clusters, fit by MM in 134 steps [CONVERGED] 
#> 
#>   Components (fit)  
#>      Tail	n = 254 (0.63) 	 Shape = 1.8 
#>   Beta C1 	n = 106 (0.24) 	 Mean = 0.5
#>   Beta C2 	n = 66 (0.13) 	 Mean = 0.23
#> 
#>   Scores (model selection)  
#>        NLL       BIC       AIC  entropy       ICL reduced.entropy     reICL
#>  -491.8202 -929.1504 -965.6404 74.25181 -854.8986       0.5409804 -928.6095
#>  size
#>     9
#> 
#> $`2`
#> [ MOBSTER ] N = 426 points with K = 3 Beta clusters, fit by MM in 493 steps [CONVERGED] 
#> 
#>   Components (fit)  
#>      Tail	n = 230 (0.54) 	 Shape = 2.1 
#>   Beta C1 	n = 12 (0.03) 	 Mean = 0.55
#>   Beta C2 	n = 84 (0.19) 	 Mean = 0.49
#>   Beta C3 	n = 100 (0.24) 	 Mean = 0.18
#> 
#>   Scores (model selection)  
#>       NLL       BIC       AIC entropy       ICL reduced.entropy     reICL size
#>  -544.266 -1015.879 -1064.532 139.062 -876.8167        4.863643 -1011.015   12
#> 
#> $`3`
#> [ MOBSTER ] N = 426 points with K = 2 Beta clusters, fit by MM in 159 steps [CONVERGED] 
#> 
#>   Components (fit)  
#>      Tail	n = 279 (0.68) 	 Shape = 1.6 
#>   Beta C1 	n = 112 (0.25) 	 Mean = 0.49
#>   Beta C2 	n = 35 (0.07) 	 Mean = 0.27
#> 
#>   Scores (model selection)  
#>        NLL       BIC       AIC  entropy       ICL reduced.entropy     reICL
#>  -487.0769 -919.6639 -956.1539 57.83922 -861.8247       0.6626803 -919.0013
#>  size
#>     9
#> 
#> $`4`
#> [ MOBSTER ] N = 426 points with K = 2 Beta clusters, fit by MM in 234 steps [CONVERGED] 
#> 
#>   Components (fit)  
#>      Tail	n = 232 (0.54) 	 Shape = 1.9 
#>   Beta C1 	n = 114 (0.26) 	 Mean = 0.5
#>   Beta C2 	n = 80 (0.2) 	 Mean = 0.19
#> 
#>   Scores (model selection)  
#>        NLL       BIC       AIC  entropy       ICL reduced.entropy     reICL
#>  -467.3392 -880.1884 -916.6784 132.9965 -747.1919        7.895264 -872.2932
#>  size
#>     9
#> 
#> $`5`
#> [ MOBSTER ] N = 426 points with K = 2 Beta clusters, fit by MM in 321 steps [CONVERGED] 
#> 
#>   Components (fit)  
#>      Tail	n = 237 (0.55) 	 Shape = 1.8 
#>   Beta C1 	n = 99 (0.22) 	 Mean = 0.5
#>   Beta C2 	n = 90 (0.22) 	 Mean = 0.2
#> 
#>   Scores (model selection)  
#>        NLL       BIC       AIC  entropy       ICL reduced.entropy     reICL
#>  -463.8741 -873.2582 -909.7481 138.3993 -734.8589        6.521388 -866.7368
#>  size
#>     9
#>